Finalized Streamlit for PPO model
