+++ DRL-OD.yml
@@ -29,14 +29,28 @@ dependencies:
   - zlib=1.2.11=h7b6447c_3
   - pip:
     - absl-py==0.9.0
+    - altair==4.0.1
     - astor==0.8.1
+    - attrs==19.3.0
+    - backcall==0.1.0
+    - base58==2.0.0
+    - bleach==3.1.1
+    - blinker==1.4
+    - boto3==1.12.16
+    - botocore==1.15.16
     - box2d-py==2.3.8
     - cachetools==4.0.0
     - certifi==2019.11.28
     - cffi==1.13.2
     - chardet==3.0.4
+    - click==7.0
     - cloudpickle==1.2.2
     - cycler==0.10.0
+    - decorator==4.4.2
+    - defusedxml==0.6.0
+    - docutils==0.15.2
+    - entrypoints==0.3
+    - enum-compat==0.0.3
     - ffmpeg==1.4
     - filelock==3.0.12
     - future==0.18.2
@@ -49,42 +63,87 @@ dependencies:
     - idna==2.8
     - imageio==2.6.1
     - imageio-ffmpeg==0.3.0
+    - importlib-metadata==1.5.0
+    - ipykernel==5.1.4
+    - ipython==7.13.0
+    - ipython-genutils==0.2.0
+    - ipywidgets==7.5.1
+    - jedi==0.16.0
+    - jinja2==2.11.1
+    - jmespath==0.9.5
     - joblib==0.14.1
+    - jsonschema==3.2.0
+    - jupyter-client==6.0.0
+    - jupyter-core==4.6.3
     - keras-applications==1.0.8
     - keras-preprocessing==1.1.0
     - kiwisolver==1.1.0
     - markdown==3.1.1
+    - markupsafe==1.1.1
     - matplotlib==3.1.2
+    - mistune==0.8.4
+    - nbconvert==5.6.1
+    - nbformat==5.0.4
+    - notebook==6.0.3
     - numpy==1.18.1
     - oauthlib==3.1.0
     - opencv-python==4.1.2.30
     - opt-einsum==3.1.0
     - pandas==1.0.1
+    - pandocfilters==1.4.2
+    - parso==0.6.2
+    - pathtools==0.1.2
+    - pexpect==4.8.0
+    - pickleshare==0.7.5
     - pillow==7.0.0
     - procgen==0.9.4
+    - prometheus-client==0.7.1
+    - prompt-toolkit==3.0.4
     - protobuf==3.11.2
+    - ptyprocess==0.6.0
     - pyasn1==0.4.8
     - pyasn1-modules==0.2.8
     - pycparser==2.19
+    - pydeck==0.3.0b2
     - pyglet==1.4.10
+    - pygments==2.5.2
     - pyparsing==2.4.6
-    - python-dateutil==2.8.1
+    - pyrsistent==0.15.7
+    - python-dateutil==2.8.0
     - pytz==2019.3
+    - pyzmq==19.0.0
     - requests==2.22.0
     - requests-oauthlib==1.3.0
     - rsa==4.0
+    - s3transfer==0.3.3
     - scikit-learn==0.22.1
     - scipy==1.4.1
+    - send2trash==1.5.0
     - setuptools==45.1.0
     - six==1.14.0
     - sklearn==0.0
+    - streamlit==0.56.0
     - tensorboard==2.1.0
     - tensorflow-estimator==2.1.0
     - tensorflow-gpu==2.1.0
     - termcolor==1.1.0
+    - terminado==0.8.3
+    - testpath==0.4.4
+    - toml==0.10.0
+    - toolz==0.10.0
+    - tornado==5.1.1
     - tqdm==4.42.0
+    - traitlets==4.3.3
+    - tzlocal==2.0.0
     - urllib3==1.25.8
+    - validators==0.14.2
+    - watchdog==0.10.2
+    - wcwidth==0.1.8
+    - webencodings==0.5.1
     - werkzeug==0.16.1
     - wheel==0.34.1
+    - widgetsnbextension==3.5.1
     - wrapt==1.11.2
+    - zipp==3.1.0
 prefix: /home/brian/miniconda3/envs/DRL-OD
+
