+++ requirements.txt
@@ -3,7 +3,15 @@
 # platform: linux-64
 _libgcc_mutex=0.1=main
 absl-py=0.9.0=pypi_0
+altair=4.0.1=pypi_0
 astor=0.8.1=pypi_0
+attrs=19.3.0=pypi_0
+backcall=0.1.0=pypi_0
+base58=2.0.0=pypi_0
+bleach=3.1.1=pypi_0
+blinker=1.4=pypi_0
+boto3=1.12.16=pypi_0
+botocore=1.15.16=pypi_0
 box2d-py=2.3.8=pypi_0
 bzip2=1.0.8=h516909a_2
 ca-certificates=2019.11.28=hecc5488_0
@@ -11,8 +19,14 @@ cachetools=4.0.0=pypi_0
 certifi=2019.11.28=pypi_0
 cffi=1.13.2=pypi_0
 chardet=3.0.4=pypi_0
+click=7.0=pypi_0
 cloudpickle=1.2.2=pypi_0
 cycler=0.10.0=pypi_0
+decorator=4.4.2=pypi_0
+defusedxml=0.6.0=pypi_0
+docutils=0.15.2=pypi_0
+entrypoints=0.3=pypi_0
+enum-compat=0.0.3=pypi_0
 ffmpeg=1.4=pypi_0
 filelock=3.0.12=pypi_0
 freetype=2.10.0=he983fc9_1
@@ -29,7 +43,18 @@ h5py=2.10.0=pypi_0
 idna=2.8=pypi_0
 imageio=2.6.1=pypi_0
 imageio-ffmpeg=0.3.0=pypi_0
+importlib-metadata=1.5.0=pypi_0
+ipykernel=5.1.4=pypi_0
+ipython=7.13.0=pypi_0
+ipython-genutils=0.2.0=pypi_0
+ipywidgets=7.5.1=pypi_0
+jedi=0.16.0=pypi_0
+jinja2=2.11.1=pypi_0
+jmespath=0.9.5=pypi_0
 joblib=0.14.1=pypi_0
+jsonschema=3.2.0=pypi_0
+jupyter-client=6.0.0=pypi_0
+jupyter-core=4.6.3=pypi_0
 keras-applications=1.0.8=pypi_0
 keras-preprocessing=1.1.0=pypi_0
 kiwisolver=1.1.0=pypi_0
@@ -40,9 +65,14 @@ libiconv=1.15=h516909a_1005
 libpng=1.6.37=hed695b0_0
 libstdcxx-ng=9.1.0=hdf63c60_0
 markdown=3.1.1=pypi_0
+markupsafe=1.1.1=pypi_0
 matplotlib=3.1.2=pypi_0
+mistune=0.8.4=pypi_0
+nbconvert=5.6.1=pypi_0
+nbformat=5.0.4=pypi_0
 ncurses=6.1=he6710b0_1
 nettle=3.3=0
+notebook=6.0.3=pypi_0
 numpy=1.18.1=pypi_0
 oauthlib=3.1.0=pypi_0
 opencv-python=4.1.2.30=pypi_0
@@ -50,38 +80,66 @@ openh264=1.8.0=hdbcaa40_1000
 openssl=1.0.2u=h516909a_0
 opt-einsum=3.1.0=pypi_0
 pandas=1.0.1=pypi_0
+pandocfilters=1.4.2=pypi_0
+parso=0.6.2=pypi_0
+pathtools=0.1.2=pypi_0
+pexpect=4.8.0=pypi_0
+pickleshare=0.7.5=pypi_0
 pillow=7.0.0=pypi_0
 pip=20.0.2=py37_1
 procgen=0.9.4=pypi_0
+prometheus-client=0.7.1=pypi_0
+prompt-toolkit=3.0.4=pypi_0
 protobuf=3.11.2=pypi_0
+ptyprocess=0.6.0=pypi_0
 pyasn1=0.4.8=pypi_0
 pyasn1-modules=0.2.8=pypi_0
 pycparser=2.19=pypi_0
+pydeck=0.3.0b2=pypi_0
 pyglet=1.4.10=pypi_0
+pygments=2.5.2=pypi_0
 pyparsing=2.4.6=pypi_0
+pyrsistent=0.15.7=pypi_0
 python=3.7.0=h6e4f718_3
-python-dateutil=2.8.1=pypi_0
+python-dateutil=2.8.0=pypi_0
 pytz=2019.3=pypi_0
+pyzmq=19.0.0=pypi_0
 readline=7.0=h7b6447c_5
 requests=2.22.0=pypi_0
 requests-oauthlib=1.3.0=pypi_0
 rsa=4.0=pypi_0
+s3transfer=0.3.3=pypi_0
 scikit-learn=0.22.1=pypi_0
 scipy=1.4.1=pypi_0
+send2trash=1.5.0=pypi_0
 setuptools=45.1.0=pypi_0
 six=1.14.0=pypi_0
 sklearn=0.0=pypi_0
 sqlite=3.30.1=h7b6447c_0
+streamlit=0.56.0=pypi_0
 tensorboard=2.1.0=pypi_0
 tensorflow-estimator=2.1.0=pypi_0
 tensorflow-gpu=2.1.0=pypi_0
 termcolor=1.1.0=pypi_0
+terminado=0.8.3=pypi_0
+testpath=0.4.4=pypi_0
 tk=8.6.8=hbc83047_0
+toml=0.10.0=pypi_0
+toolz=0.10.0=pypi_0
+tornado=5.1.1=pypi_0
 tqdm=4.42.0=pypi_0
+traitlets=4.3.3=pypi_0
+tzlocal=2.0.0=pypi_0
 urllib3=1.25.8=pypi_0
+validators=0.14.2=pypi_0
+watchdog=0.10.2=pypi_0
+wcwidth=0.1.8=pypi_0
+webencodings=0.5.1=pypi_0
 werkzeug=0.16.1=pypi_0
 wheel=0.34.1=pypi_0
+widgetsnbextension=3.5.1=pypi_0
 wrapt=1.11.2=pypi_0
 x264=1!152.20180717=h14c3975_1001
 xz=5.2.4=h14c3975_4
+zipp=3.1.0=pypi_0
 zlib=1.2.11=h7b6447c_3
